requirement.txt
numpy
pandas
yfinance
alpha-vantage
finnhub-python
#TA-Lib
#zipline
pyfolio
mplfinancepoip